# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Secure Online Transaction System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/duress-pin:
    put:
      operationId: setDuressPin
      tags: [users]
      summary: Set or update the emergency duress PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDuressPinBody"
      responses:
        "200":
          description: Duress PIN set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid current PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/pin:
    put:
      operationId: updatePin
      tags: [users]
      summary: Update PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePinBody"
      responses:
        "200":
          description: PIN updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/account/{accountNumber}:
    get:
      operationId: lookupAccount
      tags: [users]
      summary: Lookup account by account number (for transfer recipient verification)
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountLookup"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/deposit:
    post:
      operationId: deposit
      tags: [transactions]
      summary: Deposit money
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Invalid PIN or amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/withdraw:
    post:
      operationId: withdraw
      tags: [transactions]
      summary: Withdraw money
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Invalid PIN, insufficient funds, or invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/transfer:
    post:
      operationId: transfer
      tags: [transactions]
      summary: Transfer money to another account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Invalid PIN, insufficient funds, or invalid recipient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/history:
    get:
      operationId: getTransactionHistory
      tags: [transactions]
      summary: Get transaction history for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistoryResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        balance:
          type: number
        status:
          type: string
          enum: [pending, active, rejected]
        address:
          type: string
        city:
          type: string
        state:
          type: string
        createdAt:
          type: string
        duressMode:
          type: boolean
        hasDuressPin:
          type: boolean
      required: [id, accountNumber, firstName, lastName, email, phone, balance, status, address, city, state, createdAt]
    AccountLookup:
      type: object
      properties:
        accountNumber:
          type: string
        fullName:
          type: string
      required: [accountNumber, fullName]
    UpdatePinBody:
      type: object
      properties:
        currentPin:
          type: string
        newPin:
          type: string
          minLength: 4
          maxLength: 6
      required: [currentPin, newPin]
    SetDuressPinBody:
      type: object
      properties:
        currentPin:
          type: string
        duressPin:
          type: string
          minLength: 4
          maxLength: 6
      required: [currentPin, duressPin]
    DepositBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        pin:
          type: string
      required: [amount, pin]
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        pin:
          type: string
      required: [amount, pin]
    TransferBody:
      type: object
      properties:
        toAccountNumber:
          type: string
        amount:
          type: number
          minimum: 1
        pin:
          type: string
      required: [toAccountNumber, amount, pin]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, transfer]
        amount:
          type: number
        fromAccountNumber:
          type: string
          nullable: true
        toAccountNumber:
          type: string
          nullable: true
        fromName:
          type: string
          nullable: true
        toName:
          type: string
          nullable: true
        status:
          type: string
          enum: [success, failed]
        createdAt:
          type: string
      required: [id, type, amount, status, createdAt]
    TransactionResponse:
      type: object
      properties:
        message:
          type: string
        transaction:
          $ref: "#/components/schemas/Transaction"
        newBalance:
          type: number
      required: [message, transaction, newBalance]
    TransactionHistoryResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required: [transactions, total]
...